% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicas.R
\name{algum}
\alias{algum}
\title{algum}
\usage{
algum(..., remover.na = FALSE)
}
\arguments{
\item{...}{zero ou mais vetores lógicos. Outros objetos de tamanho 0 são ignorados e o resto
são convertidos forçosamente para lógico, ignorando qualquer classe.}

\item{remover.na}{vetor lógico. Se verdadeiro, os valores de NA são removidos antes do cômputo do resultado.}
}
\description{
Dado um conjunto de vetores lógicos, algum  os valores são verdadeiros?
}
\details{
wrapper da função "any"
}
