% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geral.R
\name{aplicar}
\alias{aplicar}
\title{aplica funções sobre margens de "arrays (?)"}
\usage{
aplicar(X, MARGEM, FUN, ...)
}
\arguments{
\item{X}{Uma array, como uma matriz}

\item{MARGEM}{Um vetor indicanto os subscritos aos quais a função será aplicada.
Por exemplo, para uma matrix, o número 1 indica linhas, enquando o 2 indica colunas,
c(1,2) indica linhas e colunas. Quando X tiver nomes para as dimensões, pode ser um
vetor de caracteres (palavras) indicando o nome das dimensões.
#' @export}
}
\description{
retorna um vetor, arrray ou lista de valores obtidos pela aplicação de uma função
às margens de uma matriz
}
\details{
wrapper da função "print"
}
\examples{

}
