% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geral.R
\name{classe}
\alias{classe}
\title{classe}
\usage{
classe(x, ...)
}
\arguments{
\item{x}{Um objeto do R}
}
\description{
responde qual a classe do objeto
}
\details{
wrapper da função "class"
}
\examples{
x <- "eae, rapaziada"
classe(x)

}
