% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brincando_com_tabelas.R
\name{col.medias}
\alias{col.medias}
\title{Calculando médias e somas de linhas e colunas}
\usage{
col.medias(...)
}
\description{
Calcula essas métricas para tabelas ou
matrizes numéricas, para linhas ou colunas.
}
\examples{
 <- cbind(x1 = 3, x2 = c(4:1, 2:5))
col.medias(x); col.somas(x); lin.medias(x); lin.somas(x)
}
