% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brincando_com_tabelas.R
\name{cola.linhas}
\alias{cola.linhas}
\title{cola.linhas}
\usage{
cola.linhas(...)
}
\description{
Combina tabelas e tabelas,
ou tabelas e vetore, ou até vetores e vetores,
colando as linhas do segundo objeto no fim
(à direita) do primeiro objeto.
}
\details{
Gera uma tabela no fim.

Wrapper da função cbind.

@param  ... vetores ou matrizes
 @export
 @example
 m <- cola.linhas(1, 1:7)
 m
}
