% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.fator}
\alias{como.fator}
\title{como.fator}
\usage{
como.fator(x, ...)
}
\arguments{
\item{x}{Objeto a ser convertido em fator}
}
\description{
Converte um objeto em um fator
}
\details{
wrapper da função "as.factor"
}
\examples{
x <- c(1,2,3)
como.fator(x)

}
