% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.inteiro}
\alias{como.inteiro}
\title{como.inteiro}
\usage{
como.inteiro(x, ...)
}
\arguments{
\item{x}{Objeto a ser convertido em fator}
}
\description{
Converte um objeto em um inteiro
}
\details{
wrapper da função "as.integer"
}
\examples{
x <- c("1", "2", "3",)
como.inteiro(x)

}
