% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.lista}
\alias{como.lista}
\title{como.lista}
\usage{
como.lista(x, ...)
}
\arguments{
\item{x}{Um objeto para "cuspir" no console}
}
\description{
Converte um objeto em uma lista
}
\details{
wrapper da função "list"
}
\examples{
x <- c(1,2,3)
como.lista(x)

}
