% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.logico}
\alias{como.logico}
\title{como.logico}
\usage{
como.logico(x, ...)
}
\arguments{
\item{x}{Objeto a ser convertido em lógico}
}
\description{
Converte um objeto em um objeto do tipo logico
}
\details{
wrapper da função "logic"
}
\examples{
x <- c(1,2,3)
como.logico(x)

}
