% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.matriz}
\alias{como.matriz}
\title{Converte um objeto em uma matriz}
\usage{
como.matriz(x, ...)
}
\arguments{
\item{x}{Objeto a ser convertido em matriz}
}
\description{
wrapper da função "as.matrix"
}
\examples{
x <- c(1,2,3)
como.matriz(x)

}
