% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.numerico}
\alias{como.numerico}
\title{Converte um objeto em um numérico}
\usage{
como.numerico(x, ...)
}
\arguments{
\item{x}{Objeto a ser convertido em numérico}
}
\description{
wrapper da função as.numeric
}
\examples{
x <- c(1,2,3)
como.numerico (x)
}
