% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.tabela}
\alias{como.tabela}
\title{como.tabela}
\usage{
como.tabela(x, ...)
}
\arguments{
\item{x}{Objeto para ser convertido em tabela}
}
\description{
Converte um objeto para tabela ("data.frame")
}
\details{
wrapper da função "as.data.frame"
}
\examples{
x <- c(1,2,3)
como.tabela(x)

}
