% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.texto}
\alias{como.texto}
\title{como.texto}
\usage{
como.texto(x, ...)
}
\arguments{
\item{x}{Objeto para ser convertido ao tipo texto}
}
\description{
Converte um objeto para texto
}
\details{
wrapper da função "as.character"
}
\examples{
como.texto(c(1,2,3))
}
