% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipos_de_objetos.R
\name{como.vetor}
\alias{como.vetor}
\title{Converte um objeto em um vetor}
\usage{
como.vetor(x, ...)
}
\arguments{
\item{x}{Objeto a ser convertido em vetor}
}
\description{
wrapper da função as.vector
}
\examples{
x <- c(1,2,3)
como.vetor (x)
}
