% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cospe.R
\name{cospe}
\alias{cospe}
\title{cospe}
\usage{
cospe(x)
}
\arguments{
\item{x}{Um objeto para "cuspir" no terminal}
}
\description{
cospe imprime seu argumento e o retorna invisivelmente
(via invisível (x)). É uma função genérica,
o que significa que novos métodos de impressão
podem ser facilmente adicionados para novas classes.
}
\details{
wrapper da função "print"
}
\examples{
x <- "eae, rapaziada"
cospe(x)
}
