% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerador_tipos_de_objetos.R
\name{gera_tipos}
\alias{gera_tipos}
\title{wrapper da função", f_en, "}
\usage{
gera_tipos(f_en, f_pt, o_pt)
}
\arguments{
\item{x}{Objeto a ser convertido em", o_pt, "}
}
\description{
wrapper da função", f_en, "
}
