% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicas.R
\name{todos}
\alias{todos}
\title{todoa os valores são verdadieros? (TRUE)}
\usage{
todos(..., remover.na = FALSE)
}
\arguments{
\item{...}{zero ou mais vetores lógicos. Outros objetos de tamanho 0 são ignorados e o resto
são convertidos forçosamente para lógico, ignorando qualquer classe.}

\item{remover.na}{vetor lógico. Se verdadeiro, os valores de NA são removidos antes do cômputo do resultado.}
}
\description{
Dado um conjunto de vetores lógicos, todos os valores são verdadeiros?
}
\details{
wrapper da função "all"
}
